repo restructure followup

+++ level_forecast_tools/dataload.py
@@ -34,14 +34,17 @@ def generate_dtlist(start=start_def, end=end_def, count=1):
 
 
 
-def get_archive_data(start=start_def, end=end_def, tbname='temp'):
+def get_archive_data(start=start_def, end=end_def, 
+                     overwrite=True, tbname='temp'):
+    
     dtlist = generate_dtlist(start, end)
     
     dbpath = os.sep.join((os.path.dirname(__file__), 
                           'data/archive/csv_database.db'))
     csvdb = sa.create_engine('sqlite://' + dbpath, 
                              connect_args={'timeout': 15})
-    sql.execute(f'DROP TABLE IF EXISTS {tbname}', csvdb)
+    if overwrite:
+        sql.execute(f'DROP TABLE IF EXISTS {tbname}', csvdb)
     
     load_ani = cycle(list("\|/-"))
     
@@ -63,7 +66,7 @@ def get_archive_data(start=start_def, end=end_def, tbname='temp'):
 
 
 def load_archive_data(staref, start=start_def, end=end_def, 
-                      count=1, train=False, tbname='temp', 
+                      count=1, train=False, lag=2, tbname='temp', 
                       outdir='data/archive'):
     
     dbpath = os.sep.join((os.path.dirname(__file__), 
@@ -93,7 +96,7 @@ def load_archive_data(staref, start=start_def, end=end_def,
 
 
             if train:
-                valser.iloc[:,-1] = valser.iloc[:,-1].shift(15, axis=0)
+                valser.iloc[:,-1] = valser.iloc[:,-1].shift(lag, axis=0)
             
             outpath = os.sep.join((os.getcwd(), outdir, 
                                    f'valser_{lowbound[:13]}.tmp'))
